+++ src/elements/GeoChart/GeoChartConfigs/GeoChartConfigAdapter.scatterPlot.mixin.js
@@ -42,7 +42,10 @@ export default {
           mainDimension: singleScatterPlotGroupConfig.mainDimension,
           getRadius: singleScatterPlotGroupConfig.getRadius,
           getFillColor: singleScatterPlotGroupConfig.getFillColor,
+          getOpacity: singleScatterPlotGroupConfig.getOpacity,
           onDotClick: singleScatterPlotGroupConfig.onDotClick,
+          blockMouseEvents: singleScatterPlotGroupConfig.blockMouseEvents,
+          animationsDuration: singleScatterPlotGroupConfig.animationsDuration,
           tooltip: tooltipConfig,
           cssClasses: singleScatterPlotGroupConfig.cssClasses,
           groupKey: singleScatterPlotGroupConfig.groupKey