+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## 29.5.4
+
+New:
+
+- `animationsDuration` prop to have a specific animation time for `GeoScatterPlot`
+
+## 29.5.3
+
+Fix:
+
+- cursor is `not-allowed` when `GeoDropdownRegularButton` is `disabled`
+
+## 29.5.2
+
+Fix:
+
+- `cursor` is `default` when `onDotClicked` isn't defined in `GeoScatterPlot`
+
+New:
+
+- `getOpacity` prop in `GeoScatterPlot`
+- `blockMouseEvents` prop in `GeoScatterPlot`
+- `trailingItems` slot in `GeoCollapsableBox`
+
 ## 29.5.1
 
 New: